iniciando con el proyecto en el home
