+++ package.json
@@ -9,6 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "typewriter-effect": "^2.18.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
