+++ src/app/Components/home.style.css
@@ -1,0 +1,33 @@
+.App {
+  text-align: center;
+  background-color: #282c34;
+  color: white;
+}
+
+.App-logo {
+  height: 40vh;
+  overflow: hidden ;
+  pointer-events: none;
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 90vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+  position: relative;
+}
+
+.App-link {
+  color: #61dafb;
+}
+.TypeWraper{
+  display: flex;
+  width: 100vw;
+  justify-content: center;
+  align-items: baseline;
+}
