+++ src/app/page/App/App.js
@@ -1,0 +1,12 @@
+import './App.css';
+import {Home} from "../../Components/home"
+
+function App() {
+  return (
+   <>
+     <Home />
+   </>
+  );
+}
+
+export default App;
